+++ src/Sitko.Core.Blazor.AntDesign.Server/Sitko.Core.Blazor.AntDesign.Server.csproj
@@ -1,0 +1,12 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Razor">
+
+  <PropertyGroup>
+    <TargetFrameworks>netcoreapp3.1;net5.0;net6.0</TargetFrameworks>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Sitko.Core.Blazor.AntDesign\Sitko.Core.Blazor.AntDesign.csproj" />
+    <ProjectReference Include="..\Sitko.Core.Blazor.Server\Sitko.Core.Blazor.Server.csproj" />
+  </ItemGroup>
+
+</Project>
