+++ src/Sitko.Core.Blazor.AntDesign/Components/AntFormDebug.razor
@@ -1,4 +1,4 @@
-﻿@inherits Sitko.Core.App.Blazor.Components.BaseComponent
+﻿@inherits Sitko.Core.Blazor.Components.BaseComponent
 @typeparam TEntity
 
 <Divider/>
