+++ src/Sitko.Core.Blazor.AntDesign/Components/AntFormDebug.razor.cs
@@ -1,16 +1,16 @@
 ﻿using Microsoft.AspNetCore.Components;
-using Sitko.Core.App.Blazor.Forms;
 using Sitko.Core.App.Json;
+using Sitko.Core.Blazor.Forms;
 
-namespace Sitko.Core.Blazor.AntDesignComponents.Components
+namespace Sitko.Core.Blazor.AntDesignComponents.Components;
+
+public partial class AntFormDebug<TEntity> where TEntity : class, new()
 {
-    public partial class AntFormDebug<TEntity> where TEntity : class, new()
-    {
 #if NET6_0_OR_GREATER
-        [EditorRequired]
+    [EditorRequired]
 #endif
-        [Parameter] public BaseForm<TEntity> Form { get; set; } = null!;
+    [Parameter]
+    public BaseForm<TEntity> Form { get; set; } = null!;
 
-        private string EntityJson => JsonHelper.SerializeWithMetadata(Form.Entity);
-    }
+    private string EntityJson => JsonHelper.SerializeWithMetadata(Form.Entity);
 }
