+++ src/Sitko.Core.Blazor.AntDesign/Components/BaseAntListComponent.razor
@@ -1,5 +1,5 @@
 ﻿@typeparam TItem
-@inherits Sitko.Core.App.Blazor.Components.BaseComponent
+@inherits Sitko.Core.Blazor.Components.BaseComponent
 @if (ChildContent is not null)
 {
     <AntDesign.Table TItem="TItem"
