+++ src/Sitko.Core.Blazor.FileUpload/Sitko.Core.Blazor.FileUpload.csproj
@@ -5,11 +5,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Sitko.Core.App.Blazor\Sitko.Core.App.Blazor.csproj" />
-    <ProjectReference Include="..\Sitko.Core.Storage\Sitko.Core.Storage.csproj" />
+    <ProjectReference Include="..\Sitko.Core.Blazor\Sitko.Core.Blazor.csproj"/>
+    <ProjectReference Include="..\Sitko.Core.Storage\Sitko.Core.Storage.csproj"/>
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Tewr.Blazor.FileReader" Version="3.3.1.21360" />
+    <PackageReference Include="Tewr.Blazor.FileReader" Version="3.3.1.21360"/>
   </ItemGroup>
 </Project>
