+++ src/Sitko.Core.Blazor.FluentValidation/FluentValidator.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Forms;
-using Sitko.Core.App.Blazor.Components;
+using Sitko.Core.Blazor.Components;
 using Sitko.FluentValidation.Graph;
 
 namespace Sitko.Core.Blazor.FluentValidation;
