+++ src/Sitko.Core.Blazor.MudBlazor/Components/FileUpload/MudFileUpload.razor
@@ -1,5 +1,5 @@
 ﻿@namespace Sitko.Core.Blazor.MudBlazorComponents
-@inherits Sitko.Core.App.Blazor.Components.BaseComponent
+@inherits Sitko.Core.Blazor.Components.BaseComponent
 @typeparam TValue
 
 <MudGrid>
