+++ src/Sitko.Core.Blazor.MudBlazor/Components/Form/BaseMudForm.razor
@@ -1,10 +1,10 @@
 ﻿@namespace Sitko.Core.Blazor.MudBlazorComponents
 @typeparam TEntity
-@inherits Sitko.Core.App.Blazor.Forms.BaseForm<TEntity>
+@inherits Sitko.Core.Blazor.Forms.BaseForm<TEntity>
 
 <MudEditForm TEntity="TEntity" Entity="Entity" @ref="FormInstance">
     <Sitko.Core.Blazor.FluentValidation.FluentValidator/>
-    <Sitko.Core.App.Blazor.Forms.FormEditContextCatcher Form="this"/>
+    <Sitko.Core.Blazor.Forms.FormEditContextCatcher Form="this"/>
     @ChildContentFragment
     @if (Debug)
     {
