+++ src/Sitko.Core.Blazor.MudBlazor/Components/Form/BaseMudRepositoryForm.razor
@@ -2,10 +2,10 @@
 @typeparam TEntity
 @typeparam TEntityPk
 @typeparam TRepository
-@inherits Sitko.Core.App.Blazor.Forms.BaseRepositoryForm<TEntity, TEntityPk, TRepository>
+@inherits Sitko.Core.Blazor.Forms.BaseRepositoryForm<TEntity, TEntityPk, TRepository>
 <MudEditForm TEntity="TEntity" Entity="Entity" @ref="FormInstance">
     <Sitko.Core.Blazor.FluentValidation.FluentValidator/>
-    <Sitko.Core.App.Blazor.Forms.FormEditContextCatcher Form="this"/>
+    <Sitko.Core.Blazor.Forms.FormEditContextCatcher Form="this"/>
     @ChildContentFragment
     @if (Debug)
     {
