+++ src/Sitko.Core.Blazor.MudBlazor/Components/Form/MudFormDebug.razor
@@ -1,5 +1,5 @@
 ﻿@namespace Sitko.Core.Blazor.MudBlazorComponents
-@inherits Sitko.Core.App.Blazor.Components.BaseComponent
+@inherits Sitko.Core.Blazor.Components.BaseComponent
 @typeparam TEntity
 <MudDivider Class="mb-6 mt-6"></MudDivider>
 <MudExpansionPanels MultiExpansion="true">
