+++ src/Sitko.Core.Blazor.MudBlazor/Components/Layout/MudPageLayout.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Microsoft.AspNetCore.Components;
 using MudBlazor;
-using Sitko.Core.App.Blazor.Layout;
+using Sitko.Core.Blazor.Layout;
 
 namespace Sitko.Core.Blazor.MudBlazorComponents;
 
@@ -11,5 +11,8 @@ public class MudPageLayout : BasePageLayout<MudLayoutData, MudLayoutOptions>
 
     [Parameter] public List<BreadcrumbItem> Breadcrumbs { get; set; } = new();
 
-    protected override MudLayoutData GetLayoutData() => new() { Title = Title, Description = Description, Breadcrumbs = Breadcrumbs, Extra = Extra };
+    protected override MudLayoutData GetLayoutData() => new()
+    {
+        Title = Title, Description = Description, Breadcrumbs = Breadcrumbs, Extra = Extra
+    };
 }
