--- src/Sitko.Core.Blazor.MudBlazor/MudBlazorApplication.cs
@@ -1,11 +1,0 @@
-﻿using Sitko.Core.App.Blazor;
-
-namespace Sitko.Core.Blazor.MudBlazorComponents
-{
-    public class MudBlazorApplication<TStartup> : BlazorApplication<TStartup> where TStartup : MudBlazorStartup
-    {
-        public MudBlazorApplication(string[] args) : base(args)
-        {
-        }
-    }
-}
