+++ src/Sitko.Core.Blazor.MudBlazor/Sitko.Core.Blazor.MudBlazor.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Sitko.Core.App.Blazor\Sitko.Core.App.Blazor.csproj" />
+    <ProjectReference Include="..\Sitko.Core.Blazor\Sitko.Core.Blazor.csproj" />
     <ProjectReference Include="..\Sitko.Core.Blazor.FileUpload\Sitko.Core.Blazor.FileUpload.csproj" />
     <ProjectReference Include="..\Sitko.Core.Blazor.FluentValidation\Sitko.Core.Blazor.FluentValidation.csproj" />
   </ItemGroup>
