+++ src/Sitko.Core.Blazor.Server/BlazorApplication.cs
@@ -1,0 +1,13 @@
+﻿using Sitko.Blazor.ScriptInjector;
+using Sitko.Core.App.Web;
+
+namespace Sitko.Core.Blazor.Server;
+
+public abstract class BlazorApplication<TStartup> : WebApplication<TStartup> where TStartup : BlazorStartup
+{
+    protected BlazorApplication(string[] args) : base(args) =>
+        ConfigureServices(collection =>
+        {
+            collection.AddScriptInjector();
+        });
+}
