+++ src/Sitko.Core.Blazor/IsExternalInit.cs
