+++ src/Sitko.Core.Blazor/Layout/BasePageLayout.razor
@@ -1,4 +1,4 @@
 ﻿@typeparam TLayoutData
 @typeparam TLayoutOptions
-@inherits BaseComponent
+@inherits Sitko.Core.Blazor.Components.BaseComponent
 @ChildContent
