+++ src/Sitko.Core.Blazor/Sitko.Core.Blazor.csproj
@@ -9,7 +9,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Sitko.Core.App.Web\Sitko.Core.App.Web.csproj"/>
+    <ProjectReference Include="..\Sitko.Core.App\Sitko.Core.App.csproj"/>
     <ProjectReference Include="..\Sitko.Core.Repository\Sitko.Core.Repository.csproj"/>
   </ItemGroup>
 </Project>
