+++ src/Sitko.Core.Blazor/_Imports.razor
@@ -1,8 +1,6 @@
 ﻿@using System.Net.Http
 @using Microsoft.AspNetCore.Authorization
-@using Microsoft.AspNetCore.Components.Authorization
 @using Microsoft.AspNetCore.Components.Forms
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.AspNetCore.Components.Web
 @using Microsoft.JSInterop
-@using Sitko.Core.App.Blazor.Components
